+++ source/extensions/omni.isaac.orbit/docs/CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 Fixed
 ^^^^^
 
-* Fixed attribute error raised when calling the :class:`omni.isaac.orbit.envs.mdp.TerrainBasedPositionCommand` 
+* Fixed attribute error raised when calling the :class:`omni.isaac.orbit.envs.mdp.TerrainBasedPositionCommand`
   command term.
 * Added a dummy function in :class:`omni.isaac.orbit.terrain.TerrainImporter` that returns environment
   origins as terrain-aware sampled targets. This function should be implemented by child classes based on